change type of score int -> float

+++ internal/app/db/result.go
@@ -6,6 +6,6 @@ type Result struct {
 	Id     openapi_types.UUID `db:"id"`
 	TeamId string             `db:"team_id"`
 	GameId string             `db:"game_id"`
-	Score  int                `db:"score"`
+	Score  float64            `db:"score"`
 	Rank   int                `db:"rank"`
 }
